Stick to cpp convention
Rename all h files in hpp files. To avoid
"mutiple function declaration" I put all function definition
in cpp file. It will also be clearer for a third party
person to only look at the documentation in function declaration.
