+++ include/args/fblArgs.hpp
@@ -1,0 +1,12 @@
+#ifndef FBL_ARGS_HPP
+#define FBL_ARGS_HPP
+
+#include "args/args.hpp"
+
+class FblArgs : public Args {
+public:
+  FblArgs();
+  void setArgs();
+};
+
+#endif /* end of include guard: FBL_ARGS_HPP */
