+++ include/args/fdogArgs.hpp
@@ -1,0 +1,12 @@
+#ifndef FDOG_ARGS_HPP
+#define FDOG_ARGS_HPP
+
+#include "args/args.hpp"
+
+class FdogArgs : public Args {
+public:
+  FdogArgs();
+  void setArgs();
+};
+
+#endif /* end of include guard: FDOG_ARGS_HPP */
