+++ include/args/kMeanArgs.hpp
@@ -1,0 +1,12 @@
+#ifndef KMEAN_ARGS_HPP
+#define KMEAN_ARGS_HPP
+
+#include "args/args.hpp"
+
+class KMeanArgs : public Args {
+public:
+  KMeanArgs();
+  void setArgs();
+};
+
+#endif /* end of include guard: KMEAN_ARGS_HPP */
