+++ include/bilateral.hpp
@@ -6,16 +6,6 @@
 
 using namespace cv;
 
-Mat bilateral(Mat & src) {
-  Mat result = src.clone();
-  Mat temp = src.clone();
-  for (int i(1);i < 21;i+=2) {
-    temp = result.clone();
-    bilateralFilter(temp, result, i, i*2, i/2);
-  }
-  cv::namedWindow("Abstraction", WINDOW_NORMAL);
-  cv::imshow("Abstraction", result);
-  return result;
-}
+Mat bilateral(Mat & src);
 
 #endif /* end of include guard: BILATERAL_HPP */
